+++ src/scippnexus/nxfermi_chopper.py
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
+# @author Simon Heybrock
+from .leaf import Leaf
+
+
+class NXfermi_chopper(Leaf):
+    """Fermi chopper information, can be read as a dict.
+    """